requirements.txt
# GenPreventiva - Dependencies
# Python 3.12+

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector Database
chromadb==0.4.22

# Google AI
google-generativeai==0.4.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Config
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Redis (optional, for caching)
redis==5.0.1

# Utilities
aiofiles==23.2.1
httpx==0.26.0
